renew component--ButtonWithPanel,PhontoCard; page-PairupBoard; App.js
