+++ src/components/P5AP_ButtonWithPanel/ButtonWithPanel.scss
@@ -1,0 +1,72 @@
+.button-with-panel {
+  position: absolute; /* 相对图片定位 */
+  top: 0;
+  left: 0;
+  z-index: 10; /* 提高 z-index，确保在图片之上 */
+
+  .userLevel {
+    background-color: white;
+    border: 2px solid white;
+    border-radius: 0.6rem;
+    display: flex;
+    padding: 0 0.6rem;
+    align-items: center;
+    gap: 0.375rem;
+    font-family: "Corben-Bold";
+    font-size: 0.6rem;
+    color: #333;
+    cursor: pointer;
+    white-space: nowrap;
+
+    .userLevel_levelIcon {
+      width: 0.875rem;
+      height: 0.875rem;
+    }
+  }
+
+  .info-panel {
+    position: absolute;
+    top: 1.5rem;
+    left: -1.25rem;
+    width: 32rem;
+    background-color: white;
+    border: 2px solid #d3d3d3;
+    border-radius: 0.375rem;
+    box-shadow: 0 3px 0.5rem rgba(0, 0, 0, 0.1);
+    padding: 0.6rem 1.1rem;
+    z-index: 100; /* 提高 z-index，确保在图片之上 */
+
+    .info-panel_title {
+      font-family: "Corben-Bold";
+      font-size: 0.8rem;
+      padding-bottom: 1rem;
+      color: #0954a3;
+      margin-bottom: 0.5rem;
+      border-bottom: 1px solid #ddd;
+    }
+
+    .info-panel_subtitle {
+      font-family: "Corben-Bold";
+      font-size: 0.6rem;
+      color: #007bff;
+      margin-bottom: 0.5rem;
+    }
+
+    .info-panel_text {
+      font-family: "Corben-Regular";
+      font-size: 0.7rem;
+      color: #333;
+      line-height: 1.2;
+    }
+
+    .close-button {
+      position: absolute;
+      top: 0.625rem;
+      right: 0.625rem;
+      background: transparent;
+      border: none;
+      font-size: 1rem;
+      cursor: pointer;
+    }
+  }
+}
