+++ src/components/P5AP_PhotoCard/photocard.scss
@@ -1,83 +1,85 @@
-@use "../../styles/partials/typography" as *;
-@use "../../styles/partials/global" as *;
-@use "../../styles/partials/mixins" as *;
-@use "../../styles/partials/variables" as *;
+
+
+
+
 
 .photocard {
-    background-color: white;
-    border-radius: 10px;
-    border: 1px solid #e0e0e0;
-    width: 100%;
-    max-width: 320px;
-  
-    aspect-ratio: 8 / 10;
-    display:flex;
-    flex-direction:column;
-    text-align: left;
+  position: relative;
+  background-color: white;
+  border-radius: 0.625rem;
+  border: 0.0625rem solid #e0e0e0;
+  width: 100%;
+  max-width: 20rem;
+  aspect-ratio: 8 / 10;
+  display: flex;
+  flex-direction: column;
+  text-align: left;
+  font-family: "Corben-Bold";
+
+  .photocard__level {
     font-family: "Corben-Bold";
-  
-    &__header {
-      position: relative;
-      flex-basis:70%;
-      overflow: hidden;
-  
-      .photocard__image {
-        width: 100%;
-        height: 100%;
-        object-fit: cover;
-          border-top-left-radius: 10px;
-        border-top-right-radius: 10px; 
-      }
-  
-      .photocard__level {
-        font-family: "Corben-Bold";
-        position: absolute;
-        top: 10px;
-        left: 10px;
-        background-color: whitesmoke;
-        padding: 2px 1rem;
-        border-radius: 15px;
-        font-size: 0.6rem;
-        color: #333;
-        font-weight: bold;
-      }
+    position: absolute;
+    top: 0.625rem;
+    left: 0.625rem;
+    background-color: white;
+    padding: 0.125 1rem;
+    border-radius: 1.2rem;
+    font-size: 0.6rem;
+    color: #333;
+    z-index: 10; /* 提高 z-index，确保在图片之上 */
+  }
+
+  &__header {
+
+    flex-basis: 70%;
+    /* 保持 overflow: hidden 来裁剪图片 */
+    overflow: hidden;
+    z-index: 1; /* 设置图片的 z-index 低一些 */
+    
+    .photocard__image {
+      width: 100%;
+      height: 100%;
+      object-fit: cover; /* 让图片保持比例裁剪 */
+      border-top-left-radius: 0.625rem;
+      border-top-right-radius: 0.625rem;
     }
-  
-    &__info {
-      padding: 1rem;
-      padding-top:0;
-  
-      .photocard__name {
-        font-family: "Corben-Bold";
-        font-size: 1rem;
-        padding-top: 0.5rem;
-        font-weight: bold;
-        color: #333;
-      }
-  
-      .photocard__role {
-        font-family: "Corben-Regular";
-        font-size: 0.9rem;
-        color: #555;
-  
-      }
-  
-      .photocard_briefcase {
+
+  }
+
+  &__info {
+    padding: 1rem;
+    padding-top: 0;
+
+    .photocard__name {
+      font-family: "Corben-Bold";
+      font-size: 1rem;
+      padding-top: 0.5rem;
+      font-weight: bold;
+      color: #333;
+    }
+
+    .photocard__role {
+      font-family: "Corben-Regular";
+      font-size: 0.9rem;
+      color: #555;
+    }
+
+    .photocard_briefcase {
+      width: 1.2rem;
+      margin-right: 8px;
+      color: #555;
+    }
+
+    .photocard__skills {
+      font-family: "Corben-Regular";
+      font-size: 0.9rem;
+      color: #777;
+
+      .photocard_idea {
         width: 1.2rem;
-        margin-right: 8px;
-        color: #555;
-      }
-   
-      .photocard__skills {
-        font-family: "Corben-Regular";
-        font-size: 0.9rem;
-        color: #777;
-  
-        .photocard_idea {
-          width: 1.2rem;
-          margin-right: 8px;
-          color: #ff9900;
-        }
+        margin-right: 0.5rem;
+        color: #ff9900;
       }
     }
   }
+}
