+++ src/mockDataForPhotocard.js
@@ -1,0 +1,10 @@
+
+const mockMatchedUser = {
+    profilePicture: 'https://s3-alpha-sig.figma.com/img/ac42/ff41/fa88b42e672301ff65e0dbf24db3866f?Expires=1729468800&Key-Pair-Id=APKAQ4GOSFWCVNEHN3O4&Signature=NMt-UgNdrJrHFAUpBqjiHLPTiZPTA7AxLyuGV0lcIpeWvy70GwdS-ztBChBVHjjFrRemE87lEMIBXS4Mh136CnycMPgdGjJ3AixBQS56q7E0hunRxi7k3F9R66TfQvkNqg4JBeSq4IcAxzra5AWWs48bLyF~qqFECv2WYX4Q-hO5zd4oBwtP~hYNg6SV0q96NBlM5UpSYBTIC2AkIpk87u8iDeWp2oIEGXCW4sIERuU1j~UQhaLRNBar-YGIGDxMGjGtL0QFDjmSUfyn7-cKfOqamzAgKND08YHGj8UYQomb8qL3HwpNl86deNHntI3ON3alC7bHAlg22QQt0mDJmA__', 
+    name: 'Kate Tanner',
+    role: 'UI Designer',
+    skills: ['Figma', 'Prototyping', 'Auto Layout', 'Component Structuring'],
+    level: 'Beginner',
+  };
+
+export default mockMatchedUser
