+++ src/pages/P5AP_PairupBoard/PairupBoard.scss
@@ -10,7 +10,7 @@
     flex-direction: column;
     font-family: "Corben-Bold";
     align-items: center;
-    padding: 32px;
+    padding: 2rem;
     background-color: #f9f9f9;
     min-height: 100vh;
   }
@@ -23,9 +23,9 @@
   }
  
   .pairup-board__actions {
-    margin-top: 24px;
+    margin-top: 1.5rem;
     display: flex;
-    gap: 16px;
+    gap: 1rem;
   }
 
 
@@ -37,7 +37,7 @@
  
   .btn {
     font-size: 0.8rem;
-    border-radius: 4px;
+    border-radius: 0.25rem;
     cursor: pointer;
     font-family: "Corben-Bold";
     width: 15rem;
